estudos sobre sets

+++ src/Cap18/Collections/Conjuntos/Contato/domain/IdadeContatoComparator.java
@@ -5,6 +5,6 @@ import java.util.Comparator;
 public class IdadeContatoComparator implements Comparator<Contato> {
     @Override
     public int compare(Contato o1, Contato o2) {
-        return Double.compare(o1.getIdade(), o2.getIdade()) ;
+        return Integer.compare(o1.getIdade(), o2.getIdade()) ;
     }
 }
